python-automations/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
supabase==2.10.0
apscheduler==3.10.4
requests==2.32.0
httpx==0.27.0
openai==1.50.0
python-dotenv==1.0.1
pydantic==2.9.0
pydantic-settings==2.5.0
tenacity==8.5.0
